Rename btcRecipientSelector
